requirements.txt
# Django核心
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# 数据库
mysqlclient==2.2.0
pymysql==1.1.0

# 数据处理
pandas==2.1.3
numpy==1.24.3

# 时间序列预测
statsmodels==0.14.0
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0

# 爬虫
scrapy==2.11.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
selenium==4.15.2

# 工具类
python-dateutil==2.8.2
pytz==2023.3

# 日志和监控
celery==5.3.4
redis==5.0.1

# 其他
python-dotenv==1.0.0
Pillow==10.1.0

